requirements-build.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements-build.in
#
aioconsole==0.5.1
    # via -r requirements.txt
aiohttp==3.8.3
    # via -r requirements.txt
aiosignal==1.3.1
    # via
    #   -r requirements.txt
    #   aiohttp
async-timeout==4.0.2
    # via
    #   -r requirements.txt
    #   aiohttp
attrs==22.2.0
    # via
    #   -r requirements.txt
    #   aiohttp
bleach==5.0.1
    # via readme-renderer
build==0.9.0
    # via
    #   -r requirements.txt
    #   pip-tools
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via
    #   -r requirements.txt
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   -r requirements.txt
    #   pip-tools
colorama==0.4.6
    # via -r requirements.txt
commonmark==0.9.1
    # via rich
docutils==0.19
    # via readme-renderer
ecdsa==0.18.0
    # via
    #   -r requirements.txt
    #   python-jose
frozenlist==1.3.3
    # via
    #   -r requirements.txt
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   -r requirements.txt
    #   requests
    #   yarl
importlib-metadata==6.0.0
    # via
    #   keyring
    #   twine
jaraco-classes==3.2.3
    # via keyring
keyring==23.13.1
    # via twine
more-itertools==9.0.0
    # via jaraco-classes
multidict==6.0.4
    # via
    #   -r requirements.txt
    #   aiohttp
    #   yarl
packaging==22.0
    # via
    #   -r requirements.txt
    #   build
pep517==0.13.0
    # via
    #   -r requirements.txt
    #   build
pip-tools==6.12.1
    # via -r requirements.txt
pkginfo==1.9.6
    # via twine
pyasn1==0.4.8
    # via
    #   -r requirements.txt
    #   python-jose
    #   rsa
pygments==2.14.0
    # via
    #   readme-renderer
    #   rich
python-dotenv==0.21.0
    # via -r requirements.txt
python-jose==3.3.0
    # via -r requirements.txt
readme-renderer==37.3
    # via twine
requests==2.28.2
    # via
    #   requests-toolbelt
    #   twine
requests-toolbelt==0.10.1
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.0.1
    # via twine
rsa==4.9
    # via
    #   -r requirements.txt
    #   python-jose
six==1.16.0
    # via
    #   -r requirements.txt
    #   bleach
    #   ecdsa
tomli==2.0.1
    # via
    #   -r requirements.txt
    #   build
    #   pep517
twine==4.0.2
    # via -r requirements-build.in
urllib3==1.26.14
    # via
    #   requests
    #   twine
webencodings==0.5.1
    # via bleach
wheel==0.38.4
    # via
    #   -r requirements.txt
    #   pip-tools
yarl==1.8.2
    # via
    #   -r requirements.txt
    #   aiohttp
zipp==3.11.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
